Move waitFor helper to common-lib
Refactor: extract waitFor utility function from subscription-wildcard test
to common-lib for reuse across test files.
